Save evening

+++ internal/lotto_number/model/number.go
@@ -1,6 +1,6 @@
 package model
 
 type Number struct {
-	Value		uint64
-	ValueType	uint16
+	Value		int
+	ValueType	int
 }
